NXT-432 Fixed broken scroll to links for guidance list pages :frog:
